% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleancoach.R
\name{cleancoach}
\alias{cleancoach}
\title{Get a clean dataframe of coaches, including key information}
\usage{
cleancoach(page = NULL, decending = NULL, offering = NULL,
  level = NULL, specialization = NULL, apikey = NULL)
}
\arguments{
\item{page}{The page number to load (required).}

\item{decending}{Whether to show descending results(default value: FALSE).}

\item{offering}{Only show coaches for a specific offering, please refer to the vignette
to find proper available offering options.}

\item{level}{Only show coaches for a specific level, please refer to the vignette for
available level categories.}

\item{specialization}{Only show coaches for a specific specialization, please refer to the
vignette to find proper available specializaiton options.}

\item{apikey}{Pass your own API key can increase the rate limit from 500 to 3600 per hour.}
}
\value{
A clean dataframe with following information:
        coach name, rating, reviews, bio, specializations, level, link and image.
}
\description{
This function is designed to return a clean dataframe filled by parsed responses
from the specific MUSE API request for certain coaches.
}
\examples{
\dontrun{
cleancoach(page = 1, offering = "30-Minute Career Q&A")
cleancoach(page = 2, level = "Master Coach", specialization = "College/New Grads")}
}
\keyword{coach,}
\keyword{dataframe}
\keyword{level,}
\keyword{offering,}
\keyword{specialization,}
