% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleancompany.R
\name{cleancompany}
\alias{cleancompany}
\title{Get a clean dataframe of companies, including key information}
\usage{
cleancompany(page = NULL, decending = FALSE, industry = NULL,
  size = NULL, location = NULL, apikey = NULL)
}
\arguments{
\item{page}{The page number to load (required).}

\item{decending}{Whether to show descending results (default value: FALSE).}

\item{industry}{The company industry to get, please refer to the vignette for
available industry categories.}

\item{size}{The company size to get, please refer to the vignette for available size choice.}

\item{location}{The office location to get.}

\item{apikey}{Pass your own API key can increase the rate limit from 500 to 3600 per hour.}
}
\value{
A clean dataframe with following information:
        company name, locations, industries, size, description, tags, twitter, link and logo.
}
\description{
This function is designed to return a clean dataframe filled by parsed responses
from the specific MUSE API request for certain companies.
}
\examples{
\dontrun{
cleancompany(page = 1)
cleancompany(page = 1, industry = "Tech")
cleancompany(page = 1, size = "Large Size", location = "New York City, NY")}
}
\keyword{company,}
\keyword{dataframe}
\keyword{industry,}
\keyword{location,}
\keyword{size,}
