% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanjob.R
\name{cleanjob}
\alias{cleanjob}
\title{Get a clean dataframe of jobs, including key information}
\usage{
cleanjob(page = NULL, decending = FALSE, company = NULL,
  category = NULL, level = NULL, location = NULL, apikey = NULL)
}
\arguments{
\item{page}{The page number to load (required).}

\item{decending}{Whether to show descending results(default value: FALSE).}

\item{company}{Only get jobs for these companies.}

\item{category}{The job category to get, please refer to the vignette for available categories.}

\item{level}{The experience level required for the job, please refer to the vignette
for available levels.}

\item{location}{The job location to get (you can include flexible/remote jobs from here).}

\item{apikey}{Pass your own API key can increase the rate limit from 500 to 3600 per hour.}
}
\value{
A clean dataframe with following information:
        job title, publication date, locations, categories, levels, tags, company name,
        link and job description.
}
\description{
This function is designed to return a clean dataframe filled by parsed responses
from the specific MUSE API request for certain jobs.
}
\examples{
\dontrun{
cleanjob(page = 1)
cleanjob(page = 1, level = "Entry Level", category = "Data Science")
cleanjob(page = 2, company = "Lyft", location = "Chicago, IL")}
}
\keyword{category,}
\keyword{companny,}
\keyword{dataframe}
\keyword{job,}
\keyword{level,}
