% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanpost.R
\name{cleanpost}
\alias{cleanpost}
\title{Get a clean dataframe of advice posts, including key information}
\usage{
cleanpost(page = NULL, decending = NULL, tag = NULL, apikey = NULL)
}
\arguments{
\item{page}{The page number to load (required).}

\item{decending}{Whether to show descending results(default value: FALSE).}

\item{tag}{Only show posts for a specific tag, please refer to the vignette to find proper available tags.}

\item{apikey}{Pass your own API key can increase the rate limit from 500 to 3600 per hour.}
}
\value{
A clean dataframe with following information:
        post name, publication date, type, excerpt, tags, author name, sponsor name and
        link.
}
\description{
This function is designed to return a clean dataframe filled by parsed responses
from the specific MUSE API request for certain advice posts.
}
\examples{
\dontrun{
cleanpost(page = 1, tag = "Tools & Skills")}
}
\keyword{advice}
\keyword{dataframe}
\keyword{post,}
\keyword{tag,}
