% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcoach.R
\name{getcoach}
\alias{getcoach}
\title{Get a list of coaches, intelligently filtered by several criteria specified by user}
\usage{
getcoach(page = NULL, decending = NULL, offering = NULL,
  level = NULL, specialization = NULL, apikey = NULL)
}
\arguments{
\item{page}{The page number to load (required).}

\item{decending}{Whether to show descending results(default value: FALSE).}

\item{offering}{Only show coaches for a specific offering, please refer to the vignette
to find proper available offering options.}

\item{level}{Only show coaches for a specific level, please refer to the vignette for
available level categories.}

\item{specialization}{Only show coaches for a specific specialization, please refer to the
vignette to find proper available specializaiton options.}

\item{apikey}{Pass your own API key can increase the rate limit from 500 to 3600 per hour.}
}
\value{
A named list with the following elements:
    \describe{
        \item{\strong{message}}{a list of messages including the number of results
            in specified pages, total number of results returned by the API}
        \item{\strong{response}}{a list of API-specific response values in JSON format,
            including coache name, bio, certifications, specializations, rating,
            number of reviews, etc. At this time, no further coercion is performed, so you
            may have to use functions from the \code{XML} package to extract
            the desired output. Or you can refer to the "cleancoach" function in this
            package to get a clean dataframe with some key information of list of coaches.}
    }
}
\description{
This function can make request THE MUSE API with "coaches" endponit to get a list
of coaches. The results it will return include the amount of coaches that meet the criteria,
coache name, bio, certifications, specializations, rating, number of reviews, etc.
}
\examples{
\dontrun{
getcoach(page = 1, offering = "30-Minute Career Q&A")
getcoach(page = 2, level = "Master Coach", specialization = "College/New Grads")}
}
\keyword{coach,}
\keyword{level,}
\keyword{offering,}
\keyword{specialization}
