% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcompany.R
\name{getcompany}
\alias{getcompany}
\title{Get a list of companies, intelligently filtered by several criteria specified by user}
\usage{
getcompany(page = NULL, decending = FALSE, industry = NULL,
  size = NULL, location = NULL, apikey = NULL)
}
\arguments{
\item{page}{The page number to load (required).}

\item{decending}{Whether to show descending results (default value: FALSE).}

\item{industry}{The company industry to get, please refer to the vignette for
available industry categories.}

\item{size}{The company size to get, please refer to the vignette for available size choice.}

\item{location}{The office location to get.}

\item{apikey}{Pass your own API key can increase the rate limit from 500 to 3600 per hour.}
}
\value{
A named list with the following elements:
    \describe{
        \item{\strong{message}}{a list of messages including the number of results
            in specified pages, total number of results returned by the API}
        \item{\strong{response}}{a list of API-specific response values in JSON format,
            including the name of the company, a brief description of the company,
            its location, related industries, its tag, the size, official website, twitter,
            images, etc. At this time, no further coercion is performed, so you
            may have to use functions from the \code{jsonlite} package to extract
            the desired output. Or you can refer to the "cleancompany" function in this
            package to get a clean dataframe with some key information of list of companies.}
    }
}
\description{
This function can make request THE MUSE API with "companies" endponit to get a list
of companies. The results it will return include the name of the company, a brief
description of the company, its location, related industries, its tag, the size,
official website, twitter, images, etc.
}
\examples{
\dontrun{
getcompany(page = 1)
getcompany(page = 1, industry = "Tech")
getcompany(page = 1, size = "Large Size", location = "New York City, NY")}
}
\keyword{company,}
\keyword{industry,}
\keyword{location}
\keyword{size,}
