% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getjob.R
\name{getjob}
\alias{getjob}
\title{Get a list of jobs, intelligently filtered by several criteria specified by user}
\usage{
getjob(page = NULL, decending = FALSE, company = NULL,
  category = NULL, level = NULL, location = NULL, apikey = NULL)
}
\arguments{
\item{page}{The page number to load (required).}

\item{decending}{Whether to show descending results(default value: FALSE).}

\item{company}{Only get jobs for these companies.}

\item{category}{The job category to get, please refer to the vignette for available categories.}

\item{level}{The experience level required for the job, please refer to the vignette
for available levels.}

\item{location}{The job location to get (you can include flexible/remote jobs from here).}

\item{apikey}{Pass your own API key can increase the rate limit from 500 to 3600 per hour.}
}
\value{
A named list with the following elements:
    \describe{
        \item{\strong{message}}{a list of messages including the number of results
            in specified pages, total number of results returned by the API}
        \item{\strong{response}}{a list of API-specific response values in JSON format,
            including the job description, the job title, job id, location,
            related company, etc. At this time, no further coercion is performed, so you
            may have to use functions from the \code{jsonlite} package to extract
            the desired output. Or you can refer to the "cleanjob" function in this
            package to get a clean dataframe with some key information of list of jobs.}
    }
}
\description{
This function can make request THE MUSE API with "jobs" endponit to get a list
of jobs. The results it will return include the amount of jobs that meet the criteria,
the job description, the job title, job id, location, related company, etc.
}
\examples{
\dontrun{
getjob(page = 1)
getjob(page = 1, level = "Entry Level", category = "Data Science")
getjob(page = 2, company = "Lyft", location = "Chicago, IL")}
}
\keyword{company,}
\keyword{job,}
\keyword{level,}
\keyword{location}
