% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpost.R
\name{getpost}
\alias{getpost}
\title{Get a list of posts, intelligently filtered by tag specified by user}
\usage{
getpost(page = NULL, decending = NULL, tag = NULL, apikey = NULL)
}
\arguments{
\item{page}{The page number to load (required).}

\item{decending}{Whether to show descending results(default value: FALSE).}

\item{tag}{Only show posts for a specific tag, please refer to the vignette to find proper available tags.}

\item{apikey}{Pass your own API key can increase the rate limit from 500 to 3600 per hour.}
}
\value{
A named list with the following elements:
    \describe{
        \item{\strong{message}}{a list of messages including the number of results
            in specified pages, total number of results returned by the API}
        \item{\strong{response}}{a list of API-specific response values in JSON format,
            including the name of the post, the excerpt, its tags,
            its author, the reference link, etc. At this time, no further coercion
            is performed, so you may have to use functions from the \code{jsonlite} package
            to extract the desired output. Or you can refer to the "cleanpost" function
            in this package to get a clean dataframe with some key information of list of posts.}
    }
}
\description{
This function can make request THE MUSE API with "posts" endponit to get a list
of advice posts. The results it will return include the amount of posts that meet the criteria,
the name of the post, the excerpt, its tags, its author, the reference link, etc.
}
\examples{
\dontrun{
getpost(page = 1, tag = "Tools & Skills")}
}
\keyword{advice}
\keyword{post,}
\keyword{tag}
